Move the log formatters to separate files
