+++ cli/lib/kontena_cli.rb
@@ -135,70 +135,18 @@ module Kontena
     return @logger if @logger
     if log_target.respond_to?(:tty?) && log_target.tty?
       logger = Logger.new(log_target)
-      #logger.datetime_format = "%M%S.%3N "
-      logger.formatter = CompactFormatter.new
+      require 'kontena/cli/log_formatters/compact'
+      logger.formatter = Kontena::Cli::LogFormatter::Compact.new
     else
       logger = Logger.new(log_target, 1, 1_048_576)
-      logger.formatter = ColorStrippingFormatter.new
+      require 'kontena/cli/log_formatters/strip_color'
+      logger.formatter = Kontena::Cli::LogFormatter::StripColor.new
     end
     logger.level = ENV["DEBUG"] ? Logger::DEBUG : Logger::INFO
     logger.progname = 'CLI'
     @logger = logger
   end
 
-  class ColorStrippingFormatter < Logger::Formatter
-    def msg2str(msg)
-      super(msg.kind_of?(String) ? msg.gsub(/\e+\[{1,2}[0-9;:?]+m/m, '') : msg)
-    end
-  end
-
-  class CompactFormatter < Logger::Formatter
-    def self.ms_since_first
-      ((Time.now.to_f - @first_log) * 1000).to_i
-    end
-
-    def self.ms_since_last
-      ((Time.now.to_f - @last_log) * 1000).to_i
-    ensure
-      @last_log = Time.now.to_f
-    end
-
-    def self.__init_timers__
-      @first_log = Time.now.to_f
-      @last_log = @first_log
-    end
-
-    __init_timers__
-
-    def colorize_severity(severity)
-      case severity[0..0]
-      when 'D' then Kontena.pastel.blue('D')
-      when 'W' then Kontena.pastel.yellow('W')
-      when 'I' then Kontena.pastel.cyan('I')
-      when 'E', 'F' then Kontena.pastel.red('E')
-      else severity[0..0]
-      end
-    end
-
-    def colorize_time
-      elapsed = self.class.ms_since_last
-      str = sprintf("%6s", "#{elapsed}ms")
-      if elapsed > 300
-        Kontena.pastel.red(str)
-      elsif elapsed > 100
-        Kontena.pastel.yellow(str)
-      else
-        str
-      end
-    end
-
-    LEFT_BRACKET = Kontena.pastel.cyan('[').freeze
-    RIGHT_BRACKET = Kontena.pastel.cyan(']').freeze
-
-    def call(severity, time, progname, msg)
-      "#{LEFT_BRACKET}#{colorize_severity(severity)} #{colorize_time} #{progname}#{RIGHT_BRACKET} #{msg2str(msg)}\n"
-    end
-  end
 end
 
 # Monkeypatching string to mimick 'colorize' gem
